+++ packages/hooks/src/index.ts
@@ -5,3 +5,4 @@ export * from "./useLocalStorage.js";
 export * from "./usePreload.js";
 export * from "./useScript.js";
 export * from "./useTheme.js";
+export * from "./useVisibility.js";